refactor: TextBubble 컴포저블 함수 분리
ChatContent 페이지에서 TextBubble컴포저블 함수를 새로운 컴포넌트로 분리
